fix modal tests
